Fixed incorrect vertical padding in LaTeX Renderer.
When a figure generated by LaTeX Render is saved, the extra newline added by the `tkinter.Text' widget results in a larger bottom padding than top padding. This bug was partially hidden by the hacky placement of text in the Matplotlib figure. This change fixes both issues by placing the text in a more meaningful way and stripping leading and trailing whitespace from the contents of the `tkinter.Text'.