Remove the background colour of matched brackets.
On Linux Mint and Windows, if the cursor is on an opening bracket, its corresponding closing bracked is highlighted by drawing a cyan-coloured block on it. The bracket itself, however, retains its original colour. But on Manjaro, the bracket is forcefully coloured dark grey. Initially, I forced this colour to a lighter shade, but I really like the behaviour of GVIM on Windows and Linux Mint more, so I am changing it back.