Reverted 316edcc883d0f976ce4a01d15d76f33f60bfceed. Sharp edges look better.
Although sharp edges are not in line with my desktop environment theme, there is no compulsion for the terminal to follow the same. I use it in dark mode, which is also not in line with the theme. Further, I also modified the terminal prompt to display the machine name in italices.