Settings to differentiate tabs from spaces.
I have configured GVIM to insert spaces when Tab is pressed. However, if I have to insert a tab, I'd like it to be distinguishable from spaces. GVIM offers built-in functionality to do so. Trailing spaces, for which I was using a highlight group, can be also be shown similarly.