Private variables hidden in interactive session.
Private variables are set in `before_command` and unset in `after_command`, so they are not accessible. However, there is a caveat: entering `echo $__busy` will show its value, because `__busy` is set *after* the user types the command but before the command is executed. But that is okay, because tab-completion will not display `$__busy`.

I have also replaced `__ready` with `__busy` and made the logical changes necessary for the same.

A small note about shell parameter expansion: `-z` seems to not require quoting the variable, but `-n` does, so I have quoted `${__busy+.}` with both for consistency.