`after_command` check: removed `__busy`, used `__begin`.
In theory, if `__busy` is set, `__begin` will also be set. However, if two keyboard interrupts are received in rapid succession (perhaps by holding down 'Ctrl' and 'C'), the first invocation of `before_command` may get interrupted after it sets `__busy` but before it sets `__begin`. When `after_command` runs, it sees that `__busy` is set, and calculates `delay`. In that calculation, since `__begin` is unset, it defaults to zero, so `delay` works out to `__end`. Hence, if (for instance) `__end` is 1687259926646, the elapsed time report hilariously says '468683 h 18 m 46 s 646 ms'. This is easily fixed by checking both `__busy` and `__begin` before calculating `delay`. But if `__busy` and `__begin` are always used together, `__busy` can be eliminated, and I can simply work with `__begin`.

It is worth noting that rapid keyboard interrupts do not produce this phenomenon on Peppermint, Mint and Manjaro on any machine I have tested; it is seen only on WSL. I think it is because setting `__window` involves a call to `getactivewindow`, which runs a Windows executable. (On Linux, `WINDOWID` is usually set, and if not, `getactivewindow` is a fast alias.)