More configurations for Readline in interactive Bash.
From a manual I found online, I set the configuration variables to such values that most help my workflow. Of note is the setting which makes programmable completion show the matched part of a word in colour (specifically, it uses the same colour as `ls` does for socket files). I used grey for this purpose, but had to make the shade a bit lighter so that it would be visible.