Replaced hourglass with centre dot.
On Peppermint, it looks ugly, presumably because some font is not installed by default. Nonetheless, even if looks pretty, as it does on Mint or openSUSE, it does not take up exactly one character block. The centre dot does. Also, in the command timer report, all trailing spaces are stripped.