Fixed `go build` completions for package name.
It is not possible to mix the options `-o default` and `-W "list of words"` of `compgen`. This results in incorrect completions if the default fallback is attempted. For instance, pressing Tab after typing
    go build /bin
puts a space after `/bin`, which is wrong: one may want to build packages inside that directory. The solution is to separate the default fallback and the custom completions, as done in this commit. I made it so that a leading hyphen results in option completion, and anything else results in fallback completion.