Unset `GDK_SCALE`. Removed `QT_LOGGING_RULES`.
When Cinnamon sets the display scale to 1.5, it scales GTK and Qt apps correctly. However, if `GDK_SCALE` is set, icons in GTK apps (e.g. menu bar icons in GVIM and sidebar page thumbnails in Xreader) are not scaled when they are started from the terminal, so they look disproportionately small compared to the rest of the widgets. Unsetting it causes the icons to be scaled correctly, so they have the correct sizes. (This is the case when starting the apps from the GUI session, because this variable is not set.)