Set terminal window title from Bash.
I finally figured out how to do this consistently. On Linux Mint, this was automatic, but when I switched to Manjaro, it wasn't. I figured out that it was because I had customised `PROMPT_COMMAND`, but never understood why Linux Mint didn't have an issue with that. With these changes,
* the title of the host Xfce Terminal tab is set to the working directory. (The window title is customised to also show 'Terminal'.)
* the title of the host Kitty window is set to the working terminal. (The tab title is customised to also show the index.)

Also made the notifier run in the same process. (If run in the background, a message is displayed about that job starting and terminating.)