Set completion colours as per manual.
The Zsh manual says that the colours used for completions are determined from a string written the same way `LS_COLORS` is, but the line of code I copied from Reddit removes the colons in `LS_COLORS`, and then appends the result to a pattern (which tells Zsh to colour the matched part of the completions grey). No idea why that was working. I have made it so that `LS_COLORS` is directly appended to the pattern, in keeping with what the manual says. Enabled coloured Tab completion for Bash as well (for consistency), and it is better than Zsh at this!