Re-enabled coloured completions in Zsh.
The inconsistency in colouring of completion items was fixed in b4125533d995fa3bf4da9e937f839600a51a452d, so enabling the default colours works consistently: when there is nothing to match, the default file colours (defined in `LS_COLORS` are used), but when something is matched, the matched part is coloured grey, and the rest remains uncoloured. I wont enable coloured completions in Bash, however, because the completions of `git` are displayed in the colour of broken symbolic links (see #23).