Re-enabled coloured completions in Bash.
Although I noted in d6b8bbf7efadd7e2d9a7977dd9b77b74306f49d4 that I would not do this because of blinking completion items, the fix was just to disable the colouring of orphans in `LS_COLORS`. That is a change I can live with, since broken symbolic links show both the link and the target, which is enough information to deduce which link is broken and which one is not. Also, it is worth noting that coloured completion is one thing in which Bash trumps Zsh hands down!